refactor(node): #782 move cluster-size boot gate into its own ClusterSizeGate file
